[AS-606] move deploy to layer folders
